+++ .eslintignore
@@ -1,3 +1,4 @@
+app/node_modules/**
 lib/**
 pkg/**
 dist/**
