+++ .gitignore
@@ -4,7 +4,7 @@ node_modules
 dist
 *.sw*
 /lib
-build-config.json
+app/build-config.json
 .DS_Store
 .nyc_output/
 *~
