+++ app/ui/shared/configure-store.js
@@ -14,7 +14,7 @@ import { applyMiddleware, createStore } from 'redux';
 import createLogger from 'redux-logger';
 import thunk from './thunk';
 import * as instrument from './instrument';
-import BUILD_CONFIG from '../../../build-config';
+import BUILD_CONFIG from '../../build-config';
 
 const instrumenter = _store => next => action => {
   if (action.instrument) {
