+++ package.json
@@ -27,34 +27,38 @@
     "report": "nyc report",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "install": "node build/main.js --build-deps",
-    "postinstall": "node build/main.js --build"
+    "postinstall": "cd app && npm install && cd .."
   },
   "license": "Apache-2.0",
   "dependencies": {
     "appdirectory": "0.1.0",
-    "babel-runtime": "6.11.6",
     "backoff": "2.5.0",
     "body-parser": "1.15.2",
     "bunyan": "1.8.1",
     "colors": "1.1.2",
-    "datascript-mori": "0.16.0",
     "electron-localshortcut": "0.6.1",
     "express": "4.14.0",
     "express-validator": "2.20.8",
     "express-ws": "2.0.0",
+    "fathom-web": "1.1.2",
     "fs-extra": "0.30.0",
-    "fs-promise": "0.5.0",
     "immutable": "3.7.6",
+    "lodash": "4.15.0",
     "microtime-fast": "1.0.2",
     "morgan": "1.7.0",
-    "promise-sqlite": "1.2.1",
-    "request": "2.65.0",
+    "page-metadata-parser": "0.3.0",
+    "react": "15.3.1",
+    "react-dom": "15.3.1",
+    "react-free-style": "2.2.3",
+    "react-immutable-proptypes": "2.0.0",
+    "react-redux": "4.4.5",
+    "redux": "3.5.1",
+    "redux-immutable": "3.0.8",
+    "redux-logger": "2.6.1",
     "source-map-support": "ncalexan/node-source-map-support#fileUrls-plus",
-    "thenify-all": "1.6.0",
     "true-html-escape": "0.0.2",
     "uuid": "2.0.2",
-    "ws": "1.1.1",
-    "yargs": "5.0.0"
+    "ws": "1.1.1"
   },
   "devDependencies": {
     "babel-core": "6.14.0",
@@ -71,16 +75,17 @@
     "babel-preset-react": "6.11.1",
     "babel-preset-react-optimize": "1.0.1",
     "babel-register": "6.14.0",
+    "babel-runtime": "6.11.6",
     "check-dependencies": "1.0.1",
     "copy-webpack-plugin": "3.0.0",
     "coveralls": "2.11.12",
     "cross-env": "2.0.1",
+    "datascript-mori": "0.16.0",
     "dirsum": "0.1.1",
     "electron-builder": "5.34.1",
     "electron-download": "2.1.2",
     "electron-ipc-mock": "0.0.2",
     "electron-mocha": "3.0.5",
-    "electron-rebuild": "1.1.5",
     "enzyme": "2.4.1",
     "eslint": "3.3.1",
     "eslint-config-airbnb": "10.0.1",
@@ -90,35 +95,26 @@
     "eslint-plugin-react": "6.2.0",
     "expect": "1.20.2",
     "extract-zip": "1.5.0",
-    "fathom-web": "1.1.2",
     "fetch-mock": "5.1.1",
+    "fs-promise": "0.5.0",
     "glob": "7.0.6",
     "isomorphic-fetch": "2.2.1",
     "jsdom": "9.4.2",
     "json-loader": "0.5.4",
-    "lodash": "4.15.0",
-    "mocha": "3.0.2",
     "nyc": "8.1.0",
-    "page-metadata-parser": "0.3.0",
-    "react": "15.3.1",
+    "promise-sqlite": "1.2.1",
     "react-addons-perf": "15.3.1",
     "react-addons-test-utils": "15.3.1",
-    "react-dom": "15.3.1",
-    "react-free-style": "2.2.3",
-    "react-immutable-proptypes": "2.0.0",
-    "react-redux": "4.4.5",
-    "redux": "3.5.1",
-    "redux-immutable": "3.0.8",
-    "redux-logger": "2.6.1",
     "require-dir": "0.3.0",
     "rimraf": "2.5.4",
     "semver": "5.2.0",
     "spectron": "3.3.0",
+    "thenify-all": "1.6.0",
     "tmp": "0.0.28",
     "uglify-js": "github:mishoo/UglifyJS2#harmony",
     "webpack": "1.13.2",
-    "webpack-node-externals": "1.3.3",
-    "zip-dir": "1.0.2"
+    "zip-dir": "1.0.2",
+    "yargs": "5.0.0"
   },
   "engines": {
     "node": ">=6.1.0"
@@ -131,8 +127,10 @@
       "build/**"
     ],
     "exclude": [
+      "app/node_modules/**",
       "lib/**",
-      "test/**"
+      "test/**",
+      "dist/**"
     ],
     "extension": [
       ".jsx"
