+++ test/webdriver/test-basic.js
@@ -5,7 +5,7 @@ import os from 'os';
 import expect from 'expect';
 import Driver from '../utils/driver';
 import { HOME_PAGE } from '../../app/ui/browser/constants/ui';
-import BUILD_CONFIG from '../../build-config';
+import BUILD_CONFIG from '../../app/build-config';
 
 describe('application launch', function() {
   if (BUILD_CONFIG.development || os.platform() === 'darwin') {
